pyproject.toml
[tool.poetry]
name = "fw-gear-custom-curator"
version = "0.1.7"
description = "Custom hierarchy curation gear by INC"
authors = ["Lena Sherbakov <lesh2786@colorado.edu>"]
license = "MIT"

    [tool.poetry.dependencies]
    python = "^3.8"
    lxml = "^4.6.1"
    pandas = "^1.1.4"
    nibabel = "^3.2.0"
    Pillow = "^9"
    piexif = "^1.1.3"
    pydicom = "^2.1.1"
    pypng = "^0.0.20"
    fw-file = "^1"
    flywheel-sdk = "^15.8.0"
    backoff = "^1.11.1"
    flywheel-gear-toolkit = {version = "^0.6.1", extras = ["all"]}


    [tool.poetry.dev-dependencies]
    pytest = "^6.1.2"
    pytest-cov = "^2.10.1"
    pytest-mock = "^3.3.1"
    coverage = {extras = ["toml"], version = "^5.3"}
    pylint = "^2.6.0"
    pycodestyle = "^2.6.0"
    pydocstyle = "^5.1.1"
    black = "^22"
    isort = "^5.6.4"
    rpdb = "^0.1.6"
    remote-pdb = "^2.1.0"
    dill = "^0.3.4"
    jupyter-client = "6.1.11"
    jupyter = "^1.0.0"
    docker = "^5.0.2"


[tool.pytest.ini_options]
    minversion = "6.0"
    addopts = [
      "--cov=fw_gear_hierarchy_curator",
      "--cov-report=html",
    ]
    testpaths = [
        "tests",
    ]

[tool.black]
    exclude = '''
      /(
          \.eggs         # exclude a few common directories in the
        | \.git          # root of the project
        | \.mypy_cache
        | \.tox
        | build
        | dist
        | \.cache
      )/
    '''

[build-system]
    requires = ["poetry-core>=1.0.0"]
    build-backend = "poetry.core.masonry.api"
